funk-systemd: initial import
Funkd provides a simple systemd unix domain socket deamon example with 3
types of implementations supported and with m4 autoconf library helpers
for systemd, and with an example build environemnt provided with just
autoconf, and then also with automake.

The purpose of this example set is to help you get ramped up with
systemd with the different types of build environment requirements you
may have, and provide an m4 library to help you integrate support for
systemd autoconf. This example set also illustrates a small race
condition which could be present on many typical unix daemons, which the
sd_notify() interface addresses explicitly.

This sample set provides:

  * src/m4/systemd.m4 - LGPL m4 library for autoconf
  * src/m4/paths.m4   - LGPL m4 library for expanding @VARS@ for service target

  * Legacy unix domain socket daemon example, with race condition addressed
  * Systemd notify daemon, using shared libraries
  * Systemd notify daemon, using dynamic link loader

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>