funk-systemd: add disable target and make enable use it
This avoids some errors upon install if the service was
already installed. We obviously don't need to be doing
any of this, this is general distribution / maintenance
work but since we want to provide an example for easy
testing and loading/unloading, we want to do all the
work for you.

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>