funk-systemd: add verbose exit output describing error
Without at least a description of the return value we have
to hunt down what it means, with the error it should be
easier to debug things.

Systemd will print out messages when fprintf() is used with
SD_ERR, but in order to keep the code simple we just define
SD_ERR to nothing in the case that systemd is not being used.

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>