funk-systemd: split socket service files to change permissions
This splits up the socket required into separate files due
to the different permissions we want on them. We can therefore
avoid the chmod() call as well for the systemd case. Ideally
this should work but this doesn't right now.

The order in which you pass the sockets on the service unit file
will change the order in which systemd assigns a file descriptor
to the socket, keeping the order matters.

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>