automake: add support for generic systemd init detection
This also exposes variables and setup to allow applications
which want to install systemd units or user units even if
they don't require systemd development libraries. An example
of this is imap syncing utilities, which may use systemd
timer units to check for email regularly but which *do* not
use any systemd call in its own code.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>